Part3
For the Ride class, add the following methods: addVisitorToQueue(Visitor visitor)： Add the given visitor to the end of the visitor queue and print out the visitor's name ID、 Membership status and age information.
removeVisitorFromQueue()： Remove the visitor at the head of the queue and print out the name of the removed visitor ID、 Membership status and age information. If the queue is empty, print a prompt message indicating that the queue is empty.
printQueue()： Print out the names of all visitors in the current queue ID、 Membership status and age information. And in the partThree method of the AssignmentTwo class, first an employee rideOperaor1 and a ride ride1 are created, and then multiple visitor objects are created. Then, these visitors are added to the tourist queue by calling the addVisitorToQueue method of ride1, and a visitor in the queue is removed by calling the removeVisitorFromQueue method. Finally, the printQueue method is called to print all the visitor information in the current queue.