Part4A
A new LinkedIn List has been added to the Ride class to store visitors who have already taken the ride. This LinkedList is named rideHistory. At the same time, add the following methods to the ride class:
addVisitorToHistory(Visitor visitor)： This method is used to add tourists to the historical records of amusement facilities. After a tourist rides the amusement ride, they will be removed from the queue and added to the ride's history.
checkVisitorFromHistory(Visitor visitor)： This method is used to check whether tourists are in the historical records of the amusement facility.
numberOfVisitors()： This method is used to return the number of visitors in the historical records of the amusement facility.
printRideHistory()： This method is used to print detailed information of all visitors who have already taken the amusement ride. In this method, an Iterator is used to traverse rideHistory LinkedIn List and print the name of each visitor ID、 Membership status and age.
In addition, a new Ride object was created in the partFourA method of the Assign Two class, and tourists were added to the object's history. It checked whether a certain tourist was in the history, printed the number of tourists in the history, and printed detailed information about all tourists.