Part4B
Added Visitor Comparator Class, implemented the comparator<Visitor>interface, and defined the logic for comparing based on the age and name of visitors. The compare method of this class compares visitors based on their age, and if they are of the same age, then compares them by name. At the same time, a sortRideHistory method has been added to the Ride class for sorting tourist history records. This method uses the Collections. sort method and a custom Visitor comparator to sort rideHistory. In addition, a new Ride object was created in the partFourB method of Assign Two, and 5 visitor to visitor history records were added. Use the comparator defined by the VisitorComparator class to sort tourists. Print tourist information before and after sorting, and display the sorting results.