Part5
The Ride class has added a new method called RunOneCycle to simulate running a cycling cycle. In this method, the maximum number of visitors that can be accommodated per cycling cycle is controlled based on the maximum number of passengers maxRider, and then these visitors are removed from the queue and added to the cycling history. Finally, the numOfCycles attribute will be added to record the number of cycling cycles. Due to the setting of the maxPassenger attribute during the creation of the ride class, which has the same meaning as maxRider, maxPassenger has been changed to maxRider. In addition, a ride object named ride1 was created in the partFive method of Assign Two, which is a roller coaster with a maximum of 2 passengers. Cycle 5 times, create 5 tourist objects, and add them to the queue of ride1. Print tourist information in the queue. Then call the RunOneCycle method to simulate running a cycling cycle. Print the tourist information in the queue after one cycling cycle and the tourist information in the cycling history record.