Part6
An important part of many programs is the ability to back up data to a file and then restore it as
needed. In this section of the assignment, we will add this ability to our program.
Hint for exporting and importing data
A common way to store data in a file that needs to be imported later is to use comma-separated
values (csv). This means that we store a record on a single line, and we separate values using a
comma (,). For example, imagine an object for a class called Animal has the following information:
• species: Dog
• breed: Poodle
• colour: Brown
• name: Fido
• age: 7
You could store the Animal object in the file on a single line like: Dog,Poodle,brown,Fido,7.
When you read the file, each line in the file will contain the details for a single Animal object. You
can then use the split() method from the String class to split the line into individual values and then
use the values to create a new Animal object.