Part7 and modify about part6
Partt7: The importRideHistory (String fileName) method added to the Ride class reads passenger history records from a specified file, parses each row of records into Visitor objects, and then adds these objects to the rideHistory list. In addition, in the code of partSeven in the partSeven method, the amusement facility is created, and then the ride1.importRideHistory ("ride_history. txt") method is called to import the passenger history from the file "ride_history. txt" and store it in the rideHistory variable. Finally output the number of passengers and information.
Part6: Adjusted the exportRideHistory method in the ride class to correctly export passenger history records to a file and store passenger information in comma separated format. At the same time, add the public String toString() method in the visitor class to define how the passenger object is represented as a string. When calling visitor. toString(), a string containing passenger information will be generated, which will be written to a file. Additionally, add the passenger to the application loop statement in the partSix method of AssignTwo and replace it with specific data.